SALESDEMO-116: Migrate to release pom 10.1.0-SNAPSHOT and targeting version 3.1.0-SNAPSHOT
It was still using a pre-10.0 version of targeting.

Adapted the bootstrap content to the released version.

Disabled autoexport because it gave errors I couldn't fix.

Deleted the relevance app which is no longer needed nor supported.

Dropped the 'data' and 'elastic' module and replaced them with the new demo-support modules from targeting.

Also bumped apache httpcore and client to the latest versions.

Included new request logs and visits copied from gogreen-enterprise because the file format changed.

Generating requests for gogreen 2.0 will be the next step.